Add FileUtil class for file-related utility methods
